Fix parser for python 3.9
