+++ .github/workflows/tests.yml
@@ -4,19 +4,20 @@ on: [push, pull_request]
 
 jobs:
   test:
-    name: UTs ${{ matrix.os }}
+    name: UTs ${{ matrix.os }} ${{ matrix.python-version }}
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
+        python-version: [3.7, 3.9]
 
     steps:
       - uses: actions/checkout@v1
 
-      - name: Set up Python 3.7
+      - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v1
         with:
-          python-version: 3.7
+          python-version: ${{ matrix.python-version }}
 
       - name: Install dependencies
         run: |
@@ -52,15 +53,6 @@ jobs:
         uses: actions/setup-java@v1
         with:
           java-version: 12.x.x   
-
-      - name: Clone gauge
-        run: |
-          git clone --depth=1 https://github.com/getgauge/gauge
-
-      - name: Build gauge
-        run: |
-          cd gauge
-          go run -mod=vendor build/make.go --verbose
 
       - uses: getgauge/setup-gauge@master
         with:
