+++ build.py
@@ -24,11 +24,13 @@ DEPLOY = os.path.join(cwd, 'deploy')
 def install():
     plugin_zip = create_zip()
     call(['gauge', 'uninstall', 'python', '-v', get_version()])
-    exit_code = call(['gauge', 'install', 'python', '-f', os.path.join(BIN, plugin_zip)])
+    exit_code = call(['gauge', 'install', 'python', '-f',
+                      os.path.join(BIN, plugin_zip)])
     generate_package()
     p = os.listdir("dist")[0]
     print("Installing getgauge package using pip: \n\tpip install dist/{}".format(p))
-    call([sys.executable, "-m", "pip", "install", "dist/{}".format(p), "--upgrade", "--user"])
+    call([sys.executable, "-m", "pip", "install",
+          "dist/{}".format(p), "--upgrade", "--user"])
     sys.exit(exit_code)
 
 
@@ -36,7 +38,8 @@ def create_setup_file():
     tmpl = open("setup.tmpl", "r")
     setup = open("setup.py", "w+")
     v = get_version()
-    setup.write(tmpl.read().format(v, "{\n\t\t':python_version == \"2.7\"': ['futures']\n\t}"))
+    setup.write(tmpl.read().format(
+        v, "{\n\t\t':python_version == \"2.7\"': ['futures']\n\t}"))
     setup.close()
     tmpl.close()
 
@@ -100,7 +103,8 @@ Options:
 
 def run_tests():
     pp = "PYTHONPATH"
-    os.environ[pp] = "{0}{1}{2}".format( os.environ.get(pp), os.pathsep, os.path.abspath(os.path.curdir))
+    os.environ[pp] = "{0}{1}{2}".format(os.environ.get(
+        pp), os.pathsep, os.path.abspath(os.path.curdir))
     test_dir = os.path.join(os.path.curdir, "tests")
     exit_code = 0
     for root, _, files in os.walk(test_dir):
@@ -125,7 +129,6 @@ def main():
                 sys.exit(exit_code)
             elif sys.argv[1] == '--install':
                 install()
-
 
 
 if __name__ == '__main__':
