+++ setup.tmpl
@@ -22,6 +22,6 @@ setup(
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
     ],
-    install_requires=['redBaron', 'parso', 'debugpy', 'grpcio==1.32.0', 'protobuf>=3.5.2', 'six'],
+    install_requires=['redBaron', 'debugpy', 'grpcio==1.34.0', 'protobuf>=3.5.2'],
     extras_require={1},
 )
